Added unit tests for DarknessEvaluator (and converted floats to doubles)

+++ app/src/debug/kotlin/se/gustavkarlsson/skylight/android/settings/DebugSettings.kt
@@ -2,8 +2,8 @@ package se.gustavkarlsson.skylight.android.settings
 
 interface DebugSettings {
     val isOverrideValues: Boolean
-    val kpIndex: Float
-    val geomagLatitude: Float
-    val sunZenithAngle: Float
+    val kpIndex: Double
+    val geomagLatitude: Double
+    val sunZenithAngle: Double
     val cloudPercentage: Int
 }
